src/otel/requirements.txt
-i https://pypi.org/simple
aws-lambda-powertools==2.2.0
deprecated==1.2.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
opentelemetry-api==1.13.0; python_version >= '3.7'
opentelemetry-instrumentation==0.34b0; python_version >= '3.7'
opentelemetry-instrumentation-requests==0.34b0
opentelemetry-semantic-conventions==0.34b0; python_version >= '3.7'
opentelemetry-util-http==0.34b0; python_version >= '3.7'
setuptools==65.5.1; python_version >= '3.7'
typing-extensions==4.4.0; python_version >= '3.7'
wrapt==1.14.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
